add handlebars_helpers
